tax-integration: split tax-class Admin UI logic into composable hooks
Replace the plain fetchCommerceTaxClasses/createOrUpdateCommerceTaxClass
functions with four hooks (use-config, use-commerce-proxy-action,
use-get-commerce-tax-classes, use-upsert-commerce-tax-class) so any future
Commerce resource can be wired up the same way. Rename the backend action
from commerce-tax-classes to commerce-proxy-action to reflect that it's
already a generic Commerce REST forwarder, not tax-specific.

Add @testing-library/react and jsdom (scoped per test file) so the hooks
can be exercised with renderHook, preserving the previous test coverage.
