+++ apps/tax-integration/package.json
@@ -24,10 +24,12 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.4.0",
+    "@testing-library/react": "^16.3.2",
     "@tsconfig/bases": "^1.0.26",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@vitest/coverage-v8": "^4.0.7",
+    "jsdom": "^29.1.1",
     "typescript": "^6.0.0",
     "ultracite": "^7.0.0",
     "vitest": "^4.0.7"
